Update gn_to_cmake.py for Python 3
WriteActionForEach does a divide. In Python2, which was what was
being used when this was written, `int / int -> int`. In Python3 `int /
int -> float`. It could be possible to use `//` in Python3, but
converting to `int` does the same thing for the positive integers used
here.

Bug: skia:14349
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/706379
Commit-Queue: Ben Wagner <bungeman@google.com>

+++ gn/gn_to_cmake.py
@@ -296,7 +296,7 @@ def WriteActionForEach(out, target, project, sources, synthetic_dependencies):
   all_outputs = target.properties.get('outputs', [])
   inputs = target.properties.get('sources', [])
   # TODO: consider expanding 'output_patterns' instead.
-  outputs_per_input = len(all_outputs) / len(inputs)
+  outputs_per_input = int(len(all_outputs) / len(inputs))
   for count, source in enumerate(inputs):
     source_abs_path = project.GetAbsolutePath(source)
